requirements.txt
nltk==3.4.5
pyserini==0.8.1.0
requests==2.20.0
numpy==1.17.4
tqdm==4.39.0
pandas==0.25.3
regex==2020.1.8
scikit_learn==0.22.2.post1
torchtext==0.5.0
transformers==2.8.0
